refactor: mock dto를 구현한 dto로 수정
- 인수테스트 fixture 위치 변경
