--- app/src/test/java/shop/woowasap/mock/dto/RegisterProductRequest.java
@@ -1,8 +1,0 @@
-package shop.woowasap.mock.dto;
-
-import java.time.LocalDateTime;
-
-public record RegisterProductRequest(String name, String description, String price, int quantity,
-                                     LocalDateTime startTime, LocalDateTime endTime) {
-
-}
